refactor: isolate the activities functions in a struct
